Add notification services and persistence extraction
