+++ risk_management/web.py
@@ -22,10 +22,10 @@ from urllib.parse import quote, urljoin
 from .configuration import RealtimeConfig, load_realtime_config
 from .realtime import RealtimeDataFetcher
 from .reporting import ReportManager
-from .history import PortfolioHistoryStore
 from .api_keys import load_api_keys, save_api_keys, validate_api_key_entry
 from .snapshot_utils import build_presentable_snapshot
 from services.telemetry import Telemetry
+from services.persistence import PortfolioHistoryStore
 
 logger = logging.getLogger(__name__)
