+++ services/notifications/__init__.py
@@ -1,0 +1,10 @@
+from .telegram import send_telegram_message
+from .email import send_email_message
+from .types import NotificationError, NotificationResult
+
+__all__ = [
+    "send_email_message",
+    "send_telegram_message",
+    "NotificationError",
+    "NotificationResult",
+]
