+++ services/persistence/__init__.py
@@ -1,0 +1,3 @@
+from .history import PortfolioHistoryStore
+
+__all__ = ["PortfolioHistoryStore"]
